# Conflicts:
#	src/gps/tk/Grids.cpp

+++ src/gps/GridParameters.hpp
@@ -10,7 +10,7 @@ namespace gps
         virtual ~GridParameters() {}
 
         virtual void LoadFromId(long id) = 0;
-        virtual long getId() = 0;
+        virtual size_t getId() = 0;
         virtual DatumParameters* getDatum() = 0;
         virtual double getUnitFactorToMeter() = 0;
         virtual void putDatum(DatumParameters* datumParameters) = 0;
